* move dateformatting out of UIView * add formatting code in FormatDisplay
